requirements.txt
feedparser
python-dateutil
newspaper3k
pandas
textblob
nltk
unidecode
streamlit
scikit-learn